requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
transformers==4.41.2
torch==2.3.0 # o tensorflow si prefieres
Pillow==10.3.0
numpy==1.26.4
scikit-learn==1.5.0 # Para similitud de coseno
python-multipart==0.0.9 # Para manejo de archivos subidos